+++ src/Phan/Exception/IssueException.php
@@ -20,7 +20,11 @@ use Phan\IssueInstance;
 class IssueException extends Exception
 {
 
-    /** @var IssueInstance */
+    /**
+     * @var IssueInstance
+     * An instance of an issue that was found but can't be
+     * reported on immediately.
+     */
     private $issue_instance;
 
     /**